Use dynamic height of HC9 card

+++ app/src/main/java/com/fampay/contextualcards/data/network/response/CardGroup.kt
@@ -26,5 +26,5 @@ data class CardGroup(
     val level: Int,
 
     @SerializedName("height")
-    val height: Long,
+    val height: Int,
 )
